Removido o socket mensagens

+++ Cliente.java
@@ -17,18 +17,16 @@ public class Cliente {
 	volatile JanelaDownload download;
 	Janela1 janela;
 	Socket dados;
-	Socket mensagens;
 	InputStream is; 
 	volatile long offset = 0;
 	volatile boolean pong;
-	FileOutputStream fos;
+	FileOutputStream fos = null;
 	Thread baixar = new Thread(){
 		public void run() {
 			try {
 				pong = false;
 				try {
 					dados = new Socket(host,porta);
-					mensagens = new Socket(host, porta);
 				}catch(UnknownHostException e) {
 					janela.showDialogo("Servidor não encontrado");
 					download.dispose();
@@ -43,11 +41,11 @@ public class Cliente {
 					this.stop();
 				}
 				dados.setSoTimeout(10000);
-				mensagens.setSoTimeout(10000);
 				DataOutputStream dosDados;
 				DataInputStream disDados;
 				dosDados = new DataOutputStream(dados.getOutputStream());
 				disDados = new DataInputStream(dados.getInputStream());
+				dosDados.writeInt(1);
 				dosDados.writeUTF(caminho);
 				if(disDados.readInt()==0) {
 					janela.showDialogo("Arquivo não encontrado");
@@ -85,51 +83,6 @@ public class Cliente {
 							stop();
 						}
 					};
-					Thread RTT = new Thread() {
-						public void run() {
-							while(baixar.isAlive()) {
-								try {
-									DataOutputStream dosMensagens = new DataOutputStream(mensagens.getOutputStream());
-									dosMensagens.writeUTF("PING");
-									long tempo = System.nanoTime();
-									while(!pong) {
-										yield();
-									}
-									long tempo2 = System.nanoTime();
-									download.setRTT(((double)(tempo2  - tempo))/1000000);
-									pong = false;
-									System.out.println("Cliente RODOU PING");
-									sleep(100);
-								}catch(Exception e) {
-									System.out.println(e);
-									try {
-										fos.close();
-									} catch (IOException e1) {
-									}
-								}
-							}
-							stop();
-						}
-					};
-					Thread TrataMensagens = new Thread() {
-						public void run() {
-							try {
-								while(baixar.isAlive()) {
-									String mensagem = new DataInputStream(mensagens.getInputStream()).readUTF();
-									if(mensagem.equals("PING")) {
-										new DataOutputStream(mensagens.getOutputStream()).writeUTF("PONG");
-									}else {
-										pong = true;
-									}
-								}
-								stop();
-							}catch(Exception e) {
-								System.out.println(e);
-							}
-						}
-					};
-					TrataMensagens.start();
-					RTT.start();
 					velocidade.start();
 					is = dados.getInputStream();
 					while((byteLido = is.read(buffer))!=-1&&!this.isInterrupted()) {
@@ -153,26 +106,37 @@ public class Cliente {
 		}
 	};
  
-	public Cliente(String host, int porta, String caminho, String caminhoSalvar) {
-		this.host = host;
-		this.porta = porta;
-		download = new JanelaDownload();
-		download.setCliente(this);
-		this.caminho = caminho;
-		this.caminhoSalvar = caminhoSalvar;
-		if(new File(caminhoSalvar).exists()) {
-			new File(caminhoSalvar).delete();
+	public Cliente(String host, int porta, String caminho, String caminhoSalvar, int opcao) {
+		if(opcao == 1) {
+			this.host = host;
+			this.porta = porta;
+			download = new JanelaDownload();
+			download.setCliente(this);
+			this.caminho = caminho;
+			this.caminhoSalvar = caminhoSalvar;
+			if(new File(caminhoSalvar).exists()) {
+				new File(caminhoSalvar).delete();
+				}
+			download  = new JanelaDownload();
+			download.setCliente(this);
+			download.setVisible(true);
+			baixar.start();
+		}else {
+			try {
+				Socket dados = new Socket(host, porta);
+				new DataOutputStream(dados.getOutputStream()).writeInt(0);
+			} catch (UnknownHostException e) {
+				janela.showDialogo("Host não encontrado");
+			} catch (IOException e) {
+				janela.showDialogo("Erro de I/O");
 			}
-		download  = new JanelaDownload();
-		download.setCliente(this);
-		download.setVisible(true);
-		baixar.start();
+		}
 	}
 	public void pausar() {
 		baixar.interrupt();
 		try {
 			dados.close();
-			mensagens.close();
+			fos.close();
 			is.close();
 		}catch(Exception e) {
 			
@@ -186,7 +150,6 @@ public class Cliente {
 				pong = false;
 				try {
 					dados = new Socket(host,porta);
-					mensagens = new Socket(host, porta);
 				}catch(UnknownHostException e) {
 					janela.showDialogo("Servidor não encontrado");
 					download.trocar();
@@ -201,11 +164,11 @@ public class Cliente {
 					this.stop();
 				}
 				dados.setSoTimeout(10000);
-				mensagens.setSoTimeout(10000);
 				DataOutputStream dosDados;
 				DataInputStream disDados;
 				dosDados = new DataOutputStream(dados.getOutputStream());
 				disDados = new DataInputStream(dados.getInputStream());
+				dosDados.writeInt(1);
 				dosDados.writeUTF(caminho);
 				if(disDados.readInt()==0) {
 					janela.showDialogo("Arquivo não encontrado");
@@ -243,51 +206,6 @@ public class Cliente {
 							stop();
 						}
 					};
-					Thread RTT = new Thread() {
-						public void run() {
-							while(baixar.isAlive()) {
-								try {
-									DataOutputStream dosMensagens = new DataOutputStream(mensagens.getOutputStream());
-									dosMensagens.writeUTF("PING");
-									long tempo = System.nanoTime();
-									while(!pong) {
-										yield();
-									}
-									long tempo2 = System.nanoTime();
-									download.setRTT(((double)(tempo2  - tempo))/1000000);
-									pong = false;
-									System.out.println("Cliente RODOU PING");
-									sleep(100);
-								}catch(Exception e) {
-									System.out.println(e);
-									try {
-										fos.close();
-									} catch (IOException e1) {
-									}
-								}
-							}
-							stop();
-						}
-					};
-					Thread TrataMensagens = new Thread() {
-						public void run() {
-							try {
-								while(baixar.isAlive()) {
-									String mensagem = new DataInputStream(mensagens.getInputStream()).readUTF();
-									if(mensagem.equals("PING")) {
-										new DataOutputStream(mensagens.getOutputStream()).writeUTF("PONG");
-									}else {
-										pong = true;
-									}
-								}
-								stop();
-							}catch(Exception e) {
-								System.out.println(e);
-							}
-						}
-					};
-					TrataMensagens.start();
-					RTT.start();
 					velocidade.start();
 					is = dados.getInputStream();
 					while((byteLido = is.read(buffer))!=-1&&!this.isInterrupted()) {
@@ -305,6 +223,11 @@ public class Cliente {
 				}
 			}catch(Exception e) {
 				pausar();
+				try {
+					fos.close();
+				} catch (IOException e1) {
+					
+				}
 				download.trocar();
 				System.out.println("Erro no método cliente baixar: "+e);
 			}
@@ -316,7 +239,6 @@ public class Cliente {
 	public void cancelar() {
 		try {
 			dados.close();
-			mensagens.close();
 			baixar.interrupt();
 			fos.close();
 			download.dispose();
@@ -333,6 +255,11 @@ public class Cliente {
 	}
 	public void parar() {
 		baixar.interrupt();
+		try {
+			fos.close();
+		} catch (IOException e) {
+			
+		}
 		download.dispose();
 	}
 }
